+++ CHANGELOG.md
@@ -5,3 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.2.2] - 25.08.2022
+
+- Added support for runtime provided namespace in `EasyAttributes` trait.
+  Helps with proc-macro re-exports that wishes to change namespace of the attributes.